Rename authprovider
Now that authprovider is merely doing ldap functions (it was
already doing that), be explicit and call the package ldap.
