+++ internal/ldap/membership.go
@@ -1,11 +1,9 @@
 package ldap
 
 import (
-	"crypto/tls"
 	"fmt"
 
 	"github.com/ca-gip/kubi/internal/utils"
-	"github.com/ca-gip/kubi/pkg/types"
 	"github.com/pkg/errors"
 	"gopkg.in/ldap.v2"
 )
@@ -123,21 +121,6 @@ func getGroupsContainingUser(groupBaseDN string, userDN string) ([]*ldap.Entry, 
 	return res.Entries, nil
 }
 
-// Query LDAP with default credentials and paging parameters
-func ldapQuery(request ldap.SearchRequest) (*ldap.SearchResult, error) {
-	conn, err := ldapConnectAndBind(utils.Config.Ldap.BindDN, utils.Config.Ldap.BindPassword)
-	if err != nil {
-		return nil, err
-	}
-	defer conn.Close()
-	results, err := conn.SearchWithPaging(&request, utils.Config.Ldap.PageSize)
-	if err != nil {
-		return nil, fmt.Errorf("error searching in LDAP with request %v, %v", request, err)
-	}
-	return results, nil
-
-}
-
 // Get All groups for the cluster from LDAP
 func GetAllGroups() ([]string, error) {
 
@@ -162,101 +145,4 @@ func GetAllGroups() ([]string, error) {
 		groups = append(groups, entry.GetAttributeValue("cn"))
 	}
 	return groups, nil
-}
-
-// Authenticate a user through LDAP or LDS
-// return if bind was ok, the userDN for next usage, and error if occurred
-func AuthenticateUser(username string, password string) (types.User, error) {
-
-	// Get User Distinguished Name for Standard User
-	user, err := validateUserCredentials(utils.Config.Ldap.UserBase, username, password)
-	if err == nil {
-		return user, nil
-	}
-
-	// Now handling errors to get standard user, falling back to admin user, if
-	// config allows it
-	if len(utils.Config.Ldap.AdminUserBase) <= 0 {
-		return types.User{}, fmt.Errorf("cannot find user %s in LDAP", username)
-	}
-
-	// Retry as admin
-	user, err = validateUserCredentials(utils.Config.Ldap.AdminUserBase, username, password)
-	if err != nil {
-		return types.User{}, fmt.Errorf("cannot find admin user %s in LDAP", username)
-	}
-	return user, nil
-}
-
-// Finds an user and check if its password is correct.
-func validateUserCredentials(base string, username string, password string) (types.User, error) {
-	req := ldap.NewSearchRequest(base, ldap.ScopeWholeSubtree, ldap.NeverDerefAliases, 2, 10, false, fmt.Sprintf(utils.Config.Ldap.UserFilter, username), []string{"dn", "mail"}, nil)
-
-	conn, err := ldapConnectAndBind(utils.Config.Ldap.BindDN, utils.Config.Ldap.BindPassword)
-	if err != nil {
-		return types.User{}, err
-	}
-	defer conn.Close()
-
-	res, err := conn.SearchWithPaging(req, utils.Config.Ldap.PageSize)
-
-	switch {
-	case err != nil:
-		return types.User{}, fmt.Errorf("error searching for user %s, %w", username, err)
-	case len(res.Entries) == 0:
-		return types.User{}, fmt.Errorf("no result for the user search filter '%s'", req.Filter)
-	case len(res.Entries) > 1:
-		return types.User{}, fmt.Errorf("multiple entries found for the user search filter '%s'", req.Filter)
-	}
-
-	userDN := res.Entries[0].DN
-	mail := res.Entries[0].GetAttributeValue("mail")
-	user := types.User{
-		Username: username,
-		UserDN:   userDN,
-		Email:    mail,
-	}
-
-	_, err = ldapConnectAndBind(userDN, password)
-	if err != nil {
-		return types.User{}, fmt.Errorf("cannot authenticate user %s in LDAP", username)
-	}
-	return user, nil
-}
-
-// Connect to LDAP and bind with given credentials
-func ldapConnectAndBind(login string, password string) (*ldap.Conn, error) {
-	var (
-		err  error
-		conn *ldap.Conn
-	)
-	tlsConfig := &tls.Config{
-		ServerName:         utils.Config.Ldap.Host,
-		InsecureSkipVerify: utils.Config.Ldap.SkipTLSVerification,
-	}
-
-	if utils.Config.Ldap.UseSSL {
-		conn, err = ldap.DialTLS("tcp", fmt.Sprintf("%s:%d", utils.Config.Ldap.Host, utils.Config.Ldap.Port), tlsConfig)
-	} else {
-		conn, err = ldap.Dial("tcp", fmt.Sprintf("%s:%d", utils.Config.Ldap.Host, utils.Config.Ldap.Port))
-	}
-
-	if utils.Config.Ldap.StartTLS {
-		err = conn.StartTLS(tlsConfig)
-		if err != nil {
-			return nil, errors.Wrapf(err, "unable to setup TLS connection")
-		}
-	}
-
-	if err != nil {
-		return nil, errors.Wrapf(err, "unable to create ldap connector for %s:%d", utils.Config.Ldap.Host, utils.Config.Ldap.Port)
-	}
-
-	// Bind with BindAccount
-	err = conn.Bind(login, password)
-	if err != nil {
-		return nil, errors.WithStack(err)
-	}
-
-	return conn, nil
 }
